created training dir, moved fitter
