+++ examples/uniform_weight_fitter_example.py
@@ -2,7 +2,7 @@ import sqlite3
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
-from graphnet.models.weight_fitting import UniformWeightFitter
+from graphnet.training.weight_fitting import UniformWeightFitter
 
 
 def get_truth(db):
